fix parallel execution

+++ src/main/java/common/Common.java
@@ -1,7 +1,7 @@
 package common;
 
 public class Common {
-    public static void sleep(int milliseconds) {
+    public static void sleep(long milliseconds) {
         try {
             Thread.sleep(milliseconds);
         } catch (InterruptedException e) {
